+++ models/adverts.js
@@ -1,4 +1,5 @@
 const { Schema, model } = require("mongoose");
+const Joi = require("joi");
 
 const { handleMongooseError } = require("../helpers");
 
@@ -76,4 +77,27 @@ advertsSchema.post("save", handleMongooseError);
 
 const Adverts = model("adverts", advertsSchema);
 
-module.exports = Adverts;
+const addSchema = Joi.object({
+  year: Joi.number().required(),
+  make: Joi.string().required(),
+  model: Joi.string().required(),
+  type: Joi.string().required(),
+  img: Joi.string().required(),
+  description: Joi.string().required(),
+  fuelConsumption: Joi.string().required(),
+  engineSize: Joi.string().required(),
+  accessories: Joi.string().required(),
+  functionalities: Joi.string().required(),
+  rentalPrice: Joi.string().required(),
+  rentalCompany: Joi.string().required(),
+  address: Joi.string().required(),
+  rentalConditions: Joi.string().required(),
+  mileage: Joi.number().required(),
+  favorite: Joi.boolean().default(false),
+}).messages({
+  "number.base": "{{#label}} must be a number.",
+  "string.base": "{{#label}} must be a string.",
+  "any.required": "{{#label}} is required.",
+});
+
+module.exports = { Adverts, addSchema };
